pyproject.toml
[project]
name = "smart-sentry-sidecar"
version = "0.1.0"
description = "AI Inference Sidecar for Smart Sentry"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "pillow>=12.2.0",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.4",
]
